+++ examples/app_async_flutter/lib/domain/model/channel_credentials.dart
