+++ examples/app_async_flutter/lib/domain/model/gateway/async_client_gateway.dart
@@ -1,0 +1,6 @@
+import 'package:app_async_flutter/domain/model/channel_credentials.dart';
+
+abstract class AsyncClientGateway {
+  Future<ChannelCredential?> getCredentials();
+  Future<void> callBusinessUseCase(String channelRef, int delay);
+}
