+++ examples/app_async_flutter/lib/my_app.dart
@@ -1,0 +1,24 @@
+import 'package:app_async_flutter/async_client_service.dart';
+import 'package:app_async_flutter/infraestructure/driven_adapter/api_service.dart';
+import 'package:app_async_flutter/ui/pages/home_page.dart';
+import 'package:flutter/material.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+  final String title = 'Consumer Async Data Flow';
+
+  @override
+  Widget build(BuildContext context) {
+    return AsyncClientService(
+      eventListen: "businessEvent",
+      asyncClientGateway: ApiService(),
+      child: MaterialApp(
+        title: title,
+        theme: ThemeData(
+          primarySwatch: Colors.blue,
+        ),
+        home: MyHomePage(title: title),
+      ),
+    );
+  }
+}
