+++ examples/app_async_flutter/pubspec.yaml
@@ -14,9 +14,9 @@ dependencies:
     git:
       url: https://github.com/bancolombia/async-dataflow
       path: clients/client-dart
-  dotenv: ^4.0.1
   shared_preferences: ^2.0.13
   http: ^0.13.4
+  flutter_dotenv: ^5.0.2
 
 dev_dependencies:
   flutter_test:
@@ -24,3 +24,5 @@ dev_dependencies:
   flutter_lints: ^1.0.0
 flutter:
   uses-material-design: true
+  assets:
+    - .env
